requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
Flask==3.0.3
Flask-MySQLdb==2.0.0
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
# mysqlclient==2.2.6 #included in Flask-MySQLdb
packaging==24.2
PyYAML==6.0.1
Werkzeug==3.0.6
